Namespace ApiSpecHelper to Spec::Support::ApiHelper
